feat(web): 👻 Console
- All internal usage pages were moved into the `console`.
- Links were properly updated.
- References were updated.
