+++ web/app/layout.tsx
@@ -1,56 +1,18 @@
+import { PropsWithChildren } from 'react';
+import { Analytics } from '@vercel/analytics/next';
 import './global.css';
 
-import {
-  ReactQueryProvider,
-  ClusterProvider,
-  CryptoProvider,
-  LuckyBagsProvider,
-  SolanaProvider,
-  DataFeedProvider,
-} from '@/providers';
-import { UiLayout } from '@/components/ui/ui-layout';
-import { Analytics } from '@vercel/analytics/next';
-
-const { NEXT_PUBLIC_VERCEL_ENV = 'development' } = process.env;
+// TODO: Define project metadata
 export const metadata = {
   title: 'Luckyland',
   description: 'Where Prizes 🎉 and FUN 😃 never stops!',
 };
 
-const links: { label: string; path: string; program?: boolean }[] = [
-  { label: 'Account', path: '/account' },
-  { label: 'Treasury', path: '/treasure', program: true },
-  { label: 'Games', path: '/games', program: true },
-  { label: 'Lucky', path: '/lucky', program: true },
-  { label: 'Store', path: '/store', program: true },
-];
-
-if (NEXT_PUBLIC_VERCEL_ENV !== 'production')
-  links.push({ label: 'Clusters', path: '/clusters' });
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export default function RootLayout({ children }: PropsWithChildren) {
   return (
     <html lang="en">
       <body>
-        <ReactQueryProvider>
-          <ClusterProvider>
-            <CryptoProvider>
-              <LuckyBagsProvider>
-                <SolanaProvider>
-                  <DataFeedProvider>
-                    <UiLayout links={links} env={NEXT_PUBLIC_VERCEL_ENV}>
-                      {children}
-                    </UiLayout>
-                  </DataFeedProvider>
-                </SolanaProvider>
-              </LuckyBagsProvider>
-            </CryptoProvider>
-          </ClusterProvider>
-        </ReactQueryProvider>
+        {children}
         <Analytics />
       </body>
     </html>
